refactor(packages): collapse duplicate type surfaces
