+++ packages/gateway-protocol/src/frame-guards.ts
@@ -1,3 +1,4 @@
+import { isNonEmptyProtocolString, isProtocolRecord } from "./protocol-value-normalization.js";
 import type { EventFrame, ResponseFrame } from "./schema/frames.js";
 export type {
   ConnectParams,
@@ -9,23 +10,15 @@ export type {
   ResponseFrame,
 } from "./schema/frames.js";
 
-function isRecord(value: unknown): value is Record<string, unknown> {
-  return Boolean(value) && typeof value === "object" && !Array.isArray(value);
-}
-
-function isNonEmptyString(value: unknown): value is string {
-  return typeof value === "string" && value.length > 0;
-}
-
 function isNonNegativeInteger(value: unknown): value is number {
   return typeof value === "number" && Number.isInteger(value) && value >= 0;
 }
 
 function isGatewayErrorShape(value: unknown): boolean {
-  if (!isRecord(value)) {
+  if (!isProtocolRecord(value)) {
     return false;
   }
-  if (!isNonEmptyString(value.code) || !isNonEmptyString(value.message)) {
+  if (!isNonEmptyProtocolString(value.code) || !isNonEmptyProtocolString(value.message)) {
     return false;
   }
   if (value.retryable !== undefined && typeof value.retryable !== "boolean") {
@@ -37,7 +30,11 @@ function isGatewayErrorShape(value: unknown): boolean {
 // These lightweight guards validate dispatch-critical envelope fields without
 // compiling the full schemas or rejecting additive payload fields.
 export function isGatewayEventFrame(value: unknown): value is EventFrame {
-  if (!isRecord(value) || value.type !== "event" || !isNonEmptyString(value.event)) {
+  if (
+    !isProtocolRecord(value) ||
+    value.type !== "event" ||
+    !isNonEmptyProtocolString(value.event)
+  ) {
     return false;
   }
   return value.seq === undefined || isNonNegativeInteger(value.seq);
@@ -45,9 +42,9 @@ export function isGatewayEventFrame(value: unknown): value is EventFrame {
 
 export function isGatewayResponseFrame(value: unknown): value is ResponseFrame {
   if (
-    !isRecord(value) ||
+    !isProtocolRecord(value) ||
     value.type !== "res" ||
-    !isNonEmptyString(value.id) ||
+    !isNonEmptyProtocolString(value.id) ||
     typeof value.ok !== "boolean"
   ) {
     return false;
