+++ packages/gateway-protocol/src/gateway-error-details.ts
@@ -1,3 +1,5 @@
+import { asProtocolRecord } from "./protocol-value-normalization.js";
+
 /** Gateway JSON-RPC style error codes shared by clients and server handlers. */
 export const ErrorCodes = {
   /** Client has not completed account/device linking for this gateway. */
@@ -66,15 +68,9 @@ type GatewayErrorLike = {
 
 const LEGACY_MISSING_SCOPE_PATTERN = /\bmissing scope:\s*([a-z0-9._-]+)/i;
 
-function asRecord(value: unknown): Record<string, unknown> | null {
-  return value !== null && typeof value === "object" && !Array.isArray(value)
-    ? (value as Record<string, unknown>)
-    : null;
-}
-
 /** Reads validated missing-scope details from an untrusted protocol payload. */
 export function readMissingScopeErrorDetails(details: unknown): MissingScopeErrorDetails | null {
-  const record = asRecord(details);
+  const record = asProtocolRecord(details);
   if (record?.code !== GatewayErrorDetailCodes.MISSING_SCOPE) {
     return null;
   }
@@ -93,8 +89,8 @@ export function readMissingScopeErrorDetails(details: unknown): MissingScopeErro
 }
 
 export function isMcpAppViewExpiredError(error: unknown): boolean {
-  const record = asRecord(error);
-  return asRecord(record?.details)?.code === GatewayErrorDetailCodes.MCP_APP_VIEW_EXPIRED;
+  const record = asProtocolRecord(error);
+  return asProtocolRecord(record?.details)?.code === GatewayErrorDetailCodes.MCP_APP_VIEW_EXPIRED;
 }
 
 /**
@@ -102,7 +98,7 @@ export function isMcpAppViewExpiredError(error: unknown): boolean {
  * The message fallback keeps clients compatible with gateways predating structured details.
  */
 export function readMissingScopeError(error: unknown): MissingScopeErrorDetails | null {
-  const record = asRecord(error);
+  const record = asProtocolRecord(error);
   if (!record) {
     return null;
   }
