+++ packages/gateway-protocol/src/index.ts
@@ -37,7 +37,610 @@ export * from "./schema/sessions-suggestions.js";
 export * from "./migration-api.js";
 export type * from "./public-session-catalog.js";
 export * from "./validator-registry.js";
-export * from "./schema-export-registry.js";
+// Explicit schema exports keep public protocol changes reviewable.
+export {
+  isCloudWorkerPlacementState,
+  ConnectParamsSchema,
+  GatewaySuspendTaskBlockerSchema,
+  GatewaySuspendBlockerSchema,
+  GatewaySuspendPrepareBusyResultSchema,
+  GatewaySuspendPrepareParamsSchema,
+  GatewaySuspendPrepareReadyResultSchema,
+  GatewaySuspendPrepareResultSchema,
+  GatewaySuspendStatusReadyResultSchema,
+  GatewaySuspendStatusRunningResultSchema,
+  GatewaySuspendStatusParamsSchema,
+  GatewaySuspendStatusResultSchema,
+  GatewaySuspendResumeParamsSchema,
+  GatewaySuspendResumeResultSchema,
+  HooksStatusParamsSchema,
+  GATEWAY_SERVER_CAPS,
+  HelloOkSchema,
+  RequestFrameSchema,
+  ResponseFrameSchema,
+  EventFrameSchema,
+  GatewayFrameSchema,
+  PresenceEntrySchema,
+  SnapshotSchema,
+  ErrorShapeSchema,
+  GatewayErrorDetailsSchema,
+  MissingScopeErrorDetailsSchema,
+  WizardNotFoundErrorDetailsSchema,
+  WorkerAdmissionFailureReasonSchema,
+  WorkerAdmissionHandshakeSchema,
+  WorkerAdmissionResponseFrameSchema,
+  WorkerConnectRequestFrameSchema,
+  WorkerHeartbeatParamsSchema,
+  WorkerHeartbeatRequestFrameSchema,
+  WorkerHeartbeatResponseFrameSchema,
+  WorkerLiveEventSchema,
+  WorkerLiveEventErrorDetailsSchema,
+  WorkerLiveEventErrorShapeSchema,
+  WorkerLiveEventParamsSchema,
+  WorkerLiveEventRequestFrameSchema,
+  WorkerLiveEventResponseFrameSchema,
+  WorkerLiveEventResultSchema,
+  WorkerProtocolCloseReasonSchema,
+  WorkerTranscriptCommitErrorReasonSchema,
+  WorkerTranscriptCommitErrorShapeSchema,
+  WorkerTranscriptCommitParamsSchema,
+  WorkerTranscriptCommitRequestFrameSchema,
+  WorkerTranscriptCommitResponseFrameSchema,
+  WorkerTranscriptCommitResultSchema,
+  WorkerTranscriptMessageSchema,
+  WORKER_HEARTBEAT_INTERVAL_MS,
+  WORKER_LAUNCH_V2_PROTOCOL_FEATURE,
+  WORKER_LIVE_EVENT_PROTOCOL_FEATURE,
+  WORKER_PROTOCOL_FEATURES,
+  WORKER_PROTOCOL_MAX_FEATURE_LENGTH,
+  WORKER_PROTOCOL_MAX_FEATURES,
+  WORKER_PROTOCOL_MAX_FRAME_ID_LENGTH,
+  WORKER_PROTOCOL_MAX_IDENTIFIER_LENGTH,
+  WORKER_PROTOCOL_MAX_METHOD_LENGTH,
+  WORKER_PROTOCOL_MAX_PAYLOAD_BYTES,
+  WORKER_PROVIDER_REPLAY_MAX_DATA_BYTES,
+  WORKER_PROTOCOL_METHODS,
+  WORKER_RPC_SET_VERSION,
+  WORKER_TRANSCRIPT_MAX_BATCH_MESSAGES,
+  WORKER_TRANSCRIPT_MAX_CONTENT_PARTS,
+  WORKER_TRANSCRIPT_MAX_JSON_DEPTH,
+  WORKER_TRANSCRIPT_COMMIT_PROTOCOL_FEATURE,
+  EnvironmentStatusSchema,
+  WorkerEnvironmentStateSchema,
+  WorkerTunnelStatusSchema,
+  WorkerEnvironmentMetadataSchema,
+  EnvironmentSummarySchema,
+  EnvironmentsCreateParamsSchema,
+  EnvironmentsCreateResultSchema,
+  EnvironmentsDestroyParamsSchema,
+  EnvironmentsDestroyResultSchema,
+  EnvironmentsListParamsSchema,
+  EnvironmentsListResultSchema,
+  EnvironmentsStatusParamsSchema,
+  EnvironmentsStatusResultSchema,
+  WorkerDesktopObserveParamsSchema,
+  WorkerDesktopObserveResultSchema,
+  SystemInfoParamsSchema,
+  SystemInfoResultSchema,
+  StateVersionSchema,
+  AgentEventSchema,
+  ConversationListItemSchema,
+  ConversationListParamsSchema,
+  ConversationListResultSchema,
+  ConversationSendParamsSchema,
+  ConversationSendResultSchema,
+  ConversationTurnCancelParamsSchema,
+  ConversationTurnCancelResultSchema,
+  ConversationTurnParamsSchema,
+  ConversationTurnReplySchema,
+  ConversationTurnResultSchema,
+  MessageActionParamsSchema,
+  ChatEventSchema,
+  ChatRunStartupPhaseSchema,
+  ChatStatusEventSchema,
+  SendParamsSchema,
+  PollParamsSchema,
+  AgentParamsSchema,
+  AgentIdentityParamsSchema,
+  AgentIdentityResultSchema,
+  WakeParamsSchema,
+  PushTestParamsSchema,
+  PushTestResultSchema,
+  WebPushVapidPublicKeyParamsSchema,
+  WebPushSubscribeParamsSchema,
+  WebPushUnsubscribeParamsSchema,
+  WebPushTestParamsSchema,
+  NodePairListParamsSchema,
+  NodePairApproveParamsSchema,
+  NodePairRejectParamsSchema,
+  NodePairRemoveParamsSchema,
+  NodeListParamsSchema,
+  NodePluginToolDescriptorSchema,
+  NodePluginToolsUpdateParamsSchema,
+  NodeSkillDescriptorSchema,
+  NodeSkillsUpdateParamsSchema,
+  NodePendingAckParamsSchema,
+  NodeInvokeParamsSchema,
+  NodeInvokeInputEventSchema,
+  NodeInvokeProgressParamsSchema,
+  NodeEventResultSchema,
+  NodePresenceAlivePayloadSchema,
+  NodePresenceAliveReasonSchema,
+  NodePresenceActivityPayloadSchema,
+  NodePendingDrainParamsSchema,
+  NodePendingDrainResultSchema,
+  NodePendingEnqueueParamsSchema,
+  NodePendingEnqueueResultSchema,
+  SessionsListParamsSchema,
+  SessionCatalogCapabilitiesSchema,
+  SessionCatalogDescriptorSchema,
+  SessionCatalogPullRequestSummarySchema,
+  SessionCatalogSessionSchema,
+  SessionCatalogHostSchema,
+  SessionCatalogLocatorSchema,
+  SessionCatalogSchema,
+  SessionCatalogTranscriptItemSchema,
+  SessionsCatalogListParamsSchema,
+  SessionsCatalogListResultSchema,
+  SessionsCatalogHostEventSchema,
+  SessionsCatalogReadParamsSchema,
+  SessionsCatalogReadResultSchema,
+  SessionsCatalogContinueParamsSchema,
+  SessionsCatalogContinueResultSchema,
+  SessionsCatalogArchiveParamsSchema,
+  SessionsCatalogArchiveResultSchema,
+  SessionsCatalogStartTerminalParamsSchema,
+  SessionsCatalogStartTerminalResultSchema,
+  SessionsSearchHitSchema,
+  SessionsSearchParamsSchema,
+  SessionsSearchResultSchema,
+  SessionsCleanupParamsSchema,
+  SessionsPreviewParamsSchema,
+  SessionsDescribeParamsSchema,
+  SessionsResolveParamsSchema,
+  SessionFileBrowserEntrySchema,
+  SessionFileBrowserResultSchema,
+  SessionFileContentEncodingSchema,
+  SessionFileEntrySchema,
+  SessionFileKindSchema,
+  SessionFilePreviewKindSchema,
+  SessionFileRelevanceSchema,
+  SessionsFilesGetParamsSchema,
+  SessionsFilesGetResultSchema,
+  SessionsFilesSetParamsSchema,
+  SessionsFilesSetResultSchema,
+  SessionsFilesListParamsSchema,
+  SessionsFilesListResultSchema,
+  SessionsFilesRevealParamsSchema,
+  SessionsFilesRevealResultSchema,
+  SessionDiffFileSchema,
+  SessionDiffFileStatusSchema,
+  SessionsDiffParamsSchema,
+  SessionsDiffResultSchema,
+  SessionsCompactionListParamsSchema,
+  SessionsCompactionGetParamsSchema,
+  SessionsCompactionBranchParamsSchema,
+  SessionsCompactionRestoreParamsSchema,
+  SessionBranchSchema,
+  SessionRowSchema,
+  SessionsBranchesListParamsSchema,
+  SessionsBranchesListResultSchema,
+  SessionsBranchesSwitchParamsSchema,
+  SessionsBranchesSwitchResultSchema,
+  SessionsForkParamsSchema,
+  SessionsForkResultSchema,
+  SessionsRewindParamsSchema,
+  SessionsRewindResultSchema,
+  SessionPlacementStateSchema,
+  SessionPlacementSchema,
+  SESSION_OBSERVER_HEALTH_VALUES,
+  SessionCompanionExchangeSchema,
+  SessionObserverDigestSchema,
+  SessionObserverHealthSchema,
+  SessionObserverPlanProgressSchema,
+  SessionMemberAddParamsSchema,
+  SessionMemberMutationResultSchema,
+  SessionMemberRemoveParamsSchema,
+  SessionMemberSchema,
+  SessionMembersListParamsSchema,
+  SessionMembersListResultSchema,
+  SessionSharingActionSchema,
+  SessionSharingEventSchema,
+  SessionSharingIdentitySchema,
+  SessionSharingRoleSchema,
+  SessionVisibilitySchema,
+  SessionVisibilitySetParamsSchema,
+  SessionVisibilitySetResultSchema,
+  SessionsCompanionAskParamsSchema,
+  SessionsCompanionAskResultSchema,
+  SessionsCompanionResetParamsSchema,
+  SessionsCompanionResetResultSchema,
+  SessionsCompanionStateParamsSchema,
+  SessionsCompanionStateResultSchema,
+  SessionsObserverVisibilityParamsSchema,
+  SessionsObserverVisibilityResultSchema,
+  SessionWorktreeInfoSchema,
+  SessionsCreateParamsSchema,
+  SessionsCreateResultSchema,
+  SessionsDispatchParamsSchema,
+  SessionsDispatchResultSchema,
+  SessionsReclaimParamsSchema,
+  SessionsReclaimResultSchema,
+  SessionsSendParamsSchema,
+  SessionsAbortParamsSchema,
+  SESSIONS_PATCH_MANY_MAX_TARGETS,
+  SessionsPatchParamsSchema,
+  SessionsPatchMutationSchema,
+  SessionsPatchManyTargetSchema,
+  SessionsPatchManyParamsSchema,
+  SessionsPatchManyResultSchema,
+  SessionsPluginPatchParamsSchema,
+  SessionsResetParamsSchema,
+  SessionsDeleteParamsSchema,
+  SessionGroupSchema,
+  SessionsGroupsListParamsSchema,
+  SessionsGroupsListResultSchema,
+  SessionsGroupsPutParamsSchema,
+  SessionsGroupsRenameParamsSchema,
+  SessionsGroupsDeleteParamsSchema,
+  SessionsGroupsMutationResultSchema,
+  SessionsCompactParamsSchema,
+  SessionsUsageParamsSchema,
+  SessionDiscussionStateSchema,
+  SessionDiscussionInfoSchema,
+  SessionDiscussionInfoParamsSchema,
+  SessionDiscussionInfoResultSchema,
+  SessionDiscussionOpenParamsSchema,
+  SessionDiscussionOpenResultSchema,
+  ArtifactSummarySchema,
+  ArtifactsListParamsSchema,
+  ArtifactsGetParamsSchema,
+  ArtifactsDownloadParamsSchema,
+  AuditActivityAgentRunV1Schema,
+  AuditActivityEventV1Schema,
+  AuditActivityInboundMessageV1Schema,
+  AuditActivityListParamsSchema,
+  AuditActivityListResultSchema,
+  AuditActivityOutboundMessageV1Schema,
+  AuditActivityToolActionV1Schema,
+  ExecutionIdentityContextV1Schema,
+  DecisionReceiptV1Schema,
+  AuditRunIdentityPresentV1Schema,
+  AuditRunIdentityUnknownV1Schema,
+  AuditRunIdentityUnsupportedV1Schema,
+  AuditRunIdentityAmbiguousV1Schema,
+  AuditRunIdentityV1Schema,
+  AuditRunInspectParamsSchema,
+  AuditRunInspectResultSchema,
+  AuditEventSchema,
+  AuditListParamsSchema,
+  AuditListResultSchema,
+  UserProfileSchema,
+  UsersLinkEmailParamsSchema,
+  UsersLinkEmailResultSchema,
+  UsersListParamsSchema,
+  UsersListResultSchema,
+  UsersSelfParamsSchema,
+  UsersSelfResultSchema,
+  UsersSetAvatarParamsSchema,
+  UsersSetAvatarResultSchema,
+  UsersSetDisplayNameParamsSchema,
+  UsersSetDisplayNameResultSchema,
+  TaskSuggestionSchema,
+  TaskSuggestionEventSchema,
+  TaskSuggestionResolutionSchema,
+  TaskSuggestionsAcceptParamsSchema,
+  TaskSuggestionsAcceptResultSchema,
+  TaskSuggestionsCreateParamsSchema,
+  TaskSuggestionsCreateResultSchema,
+  TaskSuggestionsDismissParamsSchema,
+  TaskSuggestionsDismissResultSchema,
+  TaskSuggestionsListParamsSchema,
+  TaskSuggestionsListResultSchema,
+  TaskSummarySchema,
+  TasksListParamsSchema,
+  TasksListResultSchema,
+  TasksGetParamsSchema,
+  TasksGetResultSchema,
+  TasksCancelParamsSchema,
+  TasksCancelResultSchema,
+  TasksRecoveryParamsSchema,
+  TasksRecoveryResultSchema,
+  ConfigGetParamsSchema,
+  ConfigSetParamsSchema,
+  ConfigApplyParamsSchema,
+  ConfigPatchParamsSchema,
+  ConfigSchemaParamsSchema,
+  ConfigSchemaLookupParamsSchema,
+  ConfigSchemaResponseSchema,
+  ConfigSchemaLookupResultSchema,
+  UpdateStatusParamsSchema,
+  UpdateStatusResultSchema,
+  UpdateAvailableSchema,
+  UpdateScheduleStateSchema,
+  UpdateHoldParamsSchema,
+  UpdateHoldResultSchema,
+  SystemAgentChatParamsSchema,
+  SystemAgentChatQuestionSchema,
+  SystemAgentChatResultSchema,
+  SystemAgentChatHistoryParamsSchema,
+  SystemAgentChatHistoryResultSchema,
+  SystemAgentChatHistoryTurnSchema,
+  SystemChangeEntrySchema,
+  SystemChangeKindSchema,
+  SystemChangeSourceSchema,
+  SystemChangesListParamsSchema,
+  SystemChangesListResultSchema,
+  SystemAgentSetupDetectParamsSchema,
+  SystemAgentSetupDetectResultSchema,
+  SystemAgentSetupVerifyParamsSchema,
+  SystemAgentSetupVerifyResultSchema,
+  SystemAgentSetupActivateParamsSchema,
+  SystemAgentSetupActivateResultSchema,
+  SystemAgentSetupAuthStartParamsSchema,
+  SystemAgentSetupAuthStartResultSchema,
+  WizardStartParamsSchema,
+  WizardNextParamsSchema,
+  WizardCancelParamsSchema,
+  WizardStatusParamsSchema,
+  WizardStepSchema,
+  WizardNextResultSchema,
+  WizardStartResultSchema,
+  WizardStatusResultSchema,
+  TalkEventSchema,
+  TalkCatalogParamsSchema,
+  TalkCatalogResultSchema,
+  TalkClientCreateParamsSchema,
+  TalkClientCreateResultSchema,
+  TalkClientCloseParamsSchema,
+  TalkClientMutationResultSchema,
+  TalkAgentControlResultSchema,
+  TalkClientSteerParamsSchema,
+  TalkClientToolCallParamsSchema,
+  TalkClientToolCallResultSchema,
+  TalkClientTranscriptParamsSchema,
+  TalkConfigParamsSchema,
+  TalkConfigResultSchema,
+  TalkSessionAppendAudioParamsSchema,
+  TalkSessionAcknowledgeMarkParamsSchema,
+  TalkSessionCancelOutputParamsSchema,
+  TalkSessionCancelTurnParamsSchema,
+  TalkSessionCreateParamsSchema,
+  TalkSessionCreateResultSchema,
+  TalkSessionJoinParamsSchema,
+  TalkSessionJoinResultSchema,
+  TalkSessionTurnParamsSchema,
+  TalkSessionTurnResultSchema,
+  TalkSessionSteerParamsSchema,
+  TalkSessionSubmitToolResultParamsSchema,
+  TalkSessionCloseParamsSchema,
+  TalkSessionOkResultSchema,
+  TalkSpeakParamsSchema,
+  TalkSpeakResultSchema,
+  TtsSpeakParamsSchema,
+  TtsSpeakResultSchema,
+  ChannelsStatusParamsSchema,
+  ChannelsStatusResultSchema,
+  ChannelsPairingListParamsSchema,
+  ChannelsPairingListResultSchema,
+  ChannelsPairingApproveParamsSchema,
+  ChannelsPairingApproveResultSchema,
+  ChannelsPairingDismissParamsSchema,
+  ChannelsPairingDismissResultSchema,
+  ChannelsStartParamsSchema,
+  ChannelsStopParamsSchema,
+  ChannelsLogoutParamsSchema,
+  WebLoginStartParamsSchema,
+  WebLoginWaitParamsSchema,
+  AgentKindSchema,
+  AgentSummarySchema,
+  AgentsFileEntrySchema,
+  AgentsCreateParamsSchema,
+  AgentsCreateResultSchema,
+  AgentsUpdateParamsSchema,
+  AgentsUpdateResultSchema,
+  AgentsDeleteParamsSchema,
+  AgentsDeleteResultSchema,
+  AgentsFilesListParamsSchema,
+  AgentsFilesListResultSchema,
+  AgentsFilesGetParamsSchema,
+  AgentsFilesGetResultSchema,
+  AgentsFilesSetParamsSchema,
+  AgentsFilesSetResultSchema,
+  AgentsWorkspaceEntrySchema,
+  AgentsWorkspaceFileSchema,
+  AgentsWorkspaceListParamsSchema,
+  AgentsWorkspaceListResultSchema,
+  AgentsWorkspaceGetParamsSchema,
+  AgentsWorkspaceGetResultSchema,
+  AgentsListParamsSchema,
+  AgentsListResultSchema,
+  CommandsListParamsSchema,
+  CommandsListResultSchema,
+  PluginCatalogEntrySchema,
+  PluginCatalogInstallActionSchema,
+  PluginSearchPackageSchema,
+  PluginSearchResultEntrySchema,
+  PluginsInstallParamsSchema,
+  PluginsInstallResultSchema,
+  PluginsListParamsSchema,
+  PluginsListResultSchema,
+  PluginsRefreshParamsSchema,
+  PluginsRefreshResultSchema,
+  PluginsSearchParamsSchema,
+  PluginsSearchResultSchema,
+  PluginsSessionActionParamsSchema,
+  PluginsSessionActionResultSchema,
+  PluginsSetEnabledParamsSchema,
+  PluginsSetEnabledResultSchema,
+  PluginsUiDescriptorsParamsSchema,
+  PluginsUiDescriptorsResultSchema,
+  PluginsUninstallParamsSchema,
+  PluginsUninstallResultSchema,
+  ModelsAuthLogoutParamsSchema,
+  ModelsAuthStatusParamsSchema,
+  ModelsListParamsSchema,
+  AuthProbeStatusSchema,
+  ModelsProbeParamsSchema,
+  ModelsProbeResultSchema,
+  ModelsProbeTargetResultSchema,
+  SkillsStatusParamsSchema,
+  ToolsCatalogParamsSchema,
+  ToolsEffectiveParamsSchema,
+  ToolsInvokeParamsSchema,
+  SkillsInstallParamsSchema,
+  SkillsCuratorActionParamsSchema,
+  SkillsCuratorActionResultSchema,
+  SkillsCuratorStatusParamsSchema,
+  SkillsCuratorStatusResultSchema,
+  SkillsSearchParamsSchema,
+  SkillsSearchResultSchema,
+  SkillsDetailParamsSchema,
+  SkillsDetailResultSchema,
+  SkillsProposalsListParamsSchema,
+  SkillsProposalsListResultSchema,
+  SkillsProposalInspectParamsSchema,
+  SkillsProposalInspectResultSchema,
+  SkillsProposalCreateParamsSchema,
+  SkillsProposalUpdateParamsSchema,
+  SkillsProposalReviseParamsSchema,
+  SkillsProposalRequestRevisionParamsSchema,
+  SkillsProposalRequestRevisionResultSchema,
+  SkillsProposalActionParamsSchema,
+  SkillProposalEvaluationSchema,
+  SkillsProposalEvaluateParamsSchema,
+  SkillsProposalEvaluateResultSchema,
+  SkillProposalLifecycleEventSchema,
+  SkillsProposalEventsListParamsSchema,
+  SkillsProposalEventsListResultSchema,
+  SkillsProposalApplyResultSchema,
+  SkillsProposalRecordResultSchema,
+  SkillsSecurityVerdictsParamsSchema,
+  SkillsSecurityVerdictsResultSchema,
+  SkillsSkillCardParamsSchema,
+  SkillsSkillCardResultSchema,
+  SkillsUploadBeginParamsSchema,
+  SkillsUploadChunkParamsSchema,
+  SkillsUploadCommitParamsSchema,
+  SkillsUpdateParamsSchema,
+  CronJobSchema,
+  CronListParamsSchema,
+  CronStatusParamsSchema,
+  CronGetParamsSchema,
+  CronAddParamsSchema,
+  CronAddResultSchema,
+  CronDeclarativeAddResultSchema,
+  CronUpdateParamsSchema,
+  CronRemoveParamsSchema,
+  CronRunParamsSchema,
+  CronRunsParamsSchema,
+  CronScratchGetParamsSchema,
+  CronScratchGetResultSchema,
+  CronScratchSetParamsSchema,
+  CronScratchSetResultSchema,
+  LogsTailParamsSchema,
+  LogsTailResultSchema,
+  TerminalOpenParamsSchema,
+  TerminalOpenResultSchema,
+  TerminalInputParamsSchema,
+  TerminalResizeParamsSchema,
+  TerminalCloseParamsSchema,
+  TerminalAttachParamsSchema,
+  TerminalAttachResultSchema,
+  TerminalSessionInfoSchema,
+  TerminalListResultSchema,
+  TerminalTextParamsSchema,
+  TerminalTextResultSchema,
+  TerminalUploadParamsSchema,
+  TerminalUploadResultSchema,
+  TerminalAckResultSchema,
+  TerminalDataEventSchema,
+  TerminalExitEventSchema,
+  TerminalEventSchema,
+  isWellFormedApprovalId,
+  ApprovalKindSchema,
+  ApprovalDecisionSchema,
+  ApprovalAllowDecisionSchema,
+  ApprovalTerminalReasonSchema,
+  PluginApprovalSeveritySchema,
+  ExecApprovalPresentationSchema,
+  PluginApprovalPresentationSchema,
+  ApprovalPresentationSchema,
+  PendingApprovalSnapshotSchema,
+  AllowedApprovalSnapshotSchema,
+  DeniedApprovalSnapshotSchema,
+  ExpiredApprovalSnapshotSchema,
+  CancelledApprovalSnapshotSchema,
+  ApprovalSnapshotSchema,
+  TerminalApprovalSnapshotSchema,
+  ApprovalGetParamsSchema,
+  ApprovalGetResultSchema,
+  ApprovalHistoryParamsSchema,
+  ApprovalHistoryResultSchema,
+  ApprovalResolveParamsSchema,
+  ApprovalResolveResultSchema,
+  SessionApprovalEventSchema,
+  SessionApprovalReplaySchema,
+  ExecApprovalsGetParamsSchema,
+  ExecApprovalsSetParamsSchema,
+  ExecApprovalGetParamsSchema,
+  ExecApprovalRequestParamsSchema,
+  ExecApprovalResolveParamsSchema,
+  QuestionAnswersSchema,
+  QuestionGetParamsSchema,
+  QuestionGetResultSchema,
+  QuestionListParamsSchema,
+  QuestionListResultSchema,
+  QuestionOptionSchema,
+  QuestionRecordSchema,
+  QuestionRequestedEventSchema,
+  QuestionRequestParamsSchema,
+  QuestionRequestQuestionSchema,
+  QuestionRequestResultSchema,
+  QuestionResolvedEventSchema,
+  QuestionResolveParamsSchema,
+  QuestionResolveResultSchema,
+  QuestionSchema,
+  QuestionStatusSchema,
+  QuestionWaitAnswerParamsSchema,
+  QuestionWaitAnswerResultSchema,
+  ChatHistoryParamsSchema,
+  ChatMetadataParamsSchema,
+  ChatSendParamsSchema,
+  ChatInjectParamsSchema,
+  ChatToolTitlesParamsSchema,
+  ChatToolTitlesResultSchema,
+  UpdateRunParamsSchema,
+  TickEventSchema,
+  ShutdownEventSchema,
+  WorktreeRecordSchema,
+  WorktreesListParamsSchema,
+  WorktreesListResultSchema,
+  WorktreesCreateParamsSchema,
+  WorktreesRemoveParamsSchema,
+  WorktreesRemoveResultSchema,
+  WorktreesRestoreParamsSchema,
+  WorktreesGcParamsSchema,
+  WorktreesGcResultSchema,
+  WorktreesBranchesParamsSchema,
+  WorktreeBranchSchema,
+  WorktreeRepositoryStatusSchema,
+  WorktreesBranchesResultSchema,
+  FsDirEntrySchema,
+  FsListDirParamsSchema,
+  FsListDirResultSchema,
+  ErrorCodes,
+  buildMissingScopeErrorDetails,
+  GatewayErrorDetailCodes,
+  errorShape,
+  missingScopeErrorShape,
+} from "./schema-modules.js";
+export {
+  MIN_CLIENT_PROTOCOL_VERSION,
+  MIN_NODE_PROTOCOL_VERSION,
+  MIN_PROBE_PROTOCOL_VERSION,
+  PROTOCOL_VERSION,
+} from "./version.js";
 export type * from "./schema-types.js";
 
 // Local structural result keeps this package independent of core session types.
