+++ packages/media-understanding-common/src/provider-supports.ts
@@ -1,11 +1,17 @@
 // Media Understanding Common module implements provider supports behavior.
-import type { MediaUnderstandingCapability, MediaUnderstandingProvider } from "./types.js";
+import type { MediaUnderstandingCapability } from "./types.js";
+
+type MediaCapabilityProvider = {
+  transcribeAudio?: unknown;
+  describeImage?: unknown;
+  describeVideo?: unknown;
+};
 
 // Capability checks for media-understanding provider objects.
 
 /** Return true when a provider exposes the method for a media capability. */
 export function providerSupportsCapability(
-  provider: MediaUnderstandingProvider | undefined,
+  provider: MediaCapabilityProvider | undefined,
   capability: MediaUnderstandingCapability,
 ): boolean {
   if (!provider) {
