+++ src/acp/control-plane/manager.types.ts
@@ -7,6 +7,7 @@ import type {
   AcpRuntimePromptMode,
   AcpRuntimeSessionMode,
   AcpRuntimeStatus,
+  AcpRuntimeTurnAttachment,
 } from "@openclaw/acp-core/runtime/types";
 import type {
   SessionAcpIdentity,
@@ -54,10 +55,7 @@ export type AcpInitializeSessionInput = {
   backendId?: string;
 };
 
-export type AcpTurnAttachment = {
-  mediaType: string;
-  data: string;
-};
+export type AcpTurnAttachment = AcpRuntimeTurnAttachment;
 
 /** Input for one ACP prompt turn routed through the manager. */
 export type AcpRunTurnInput = {
